Moved package selectors and version pickers
To be consistent with DebianSelector and DebianVersionPicker

Test cases:
-tested package searching for all affected providers in the harvest page.
-SourcePicker (using GitHubSelector and GitHubCommitPicker) is also tested via definition page
(http://localhost:3000/definitions/pod/cocoapods/-/SoftButton/0.1.0) by editting Described.Source.
