+++ src/components/Providers/Selectors/ComposerSelector.js
@@ -3,19 +3,18 @@
 
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { getRubyGemsSearch } from '../api/clearlyDefined'
+import { getComposerSearch } from '../../../api/clearlyDefined'
 import { AsyncTypeahead } from 'react-bootstrap-typeahead'
-import searchSvg from '../images/icons/searchSvg.svg'
-import 'react-bootstrap-typeahead/css/Typeahead.css'
+import searchSvg from '../../../images/icons/searchSvg.svg'
 
-export default class RubyGemsSelector extends Component {
+export default class ComposerSelector extends Component {
   static propTypes = {
     onChange: PropTypes.func
   }
 
   constructor(props) {
     super(props)
-    this.state = { isLoading: false, options: [], focus: true }
+    this.state = { isLoading: false, options: [], focus: false }
     this.getOptions = this.getOptions.bind(this)
     this.onChange = this.onChange.bind(this)
   }
@@ -23,13 +22,13 @@ export default class RubyGemsSelector extends Component {
   onChange(values) {
     const { onChange } = this.props
     const value = values.length === 0 ? null : values[0]
-    value && onChange && onChange({ type: 'gem', provider: 'rubygems', name: value.id }, 'package')
+    value && onChange && onChange({ type: 'composer', provider: 'packagist', name: value.id }, 'package')
   }
 
   async getOptions(value) {
     try {
       this.setState({ ...this.state, isLoading: true })
-      const options = await getRubyGemsSearch(this.props.token, value)
+      const options = await getComposerSearch(this.props.token, value)
       this.setState({ ...this.state, options, isLoading: false })
     } catch (error) {
       this.setState({ ...this.state, options: [], isLoading: false })
@@ -44,11 +43,11 @@ export default class RubyGemsSelector extends Component {
           <img src={searchSvg} alt="search" />
         </div>
         <AsyncTypeahead
-          id="ruby-selector"
+          id="composer-selector"
           className="harvest-search"
           useCache={false}
           options={options}
-          placeholder={'Pick a RubyGem to harvest'}
+          placeholder={'Pick a Composer to harvest'}
           onChange={this.onChange}
           labelKey="id"
           onFocus={() => this.setState({ ...this.state, focus: true })}
