+++ src/components/Providers/Selectors/CrateSelector.js
@@ -1,21 +1,20 @@
-// (c) Copyright 2022, SAP SE and ClearlyDefined contributors. Licensed under the MIT license.
+// Copyright (c) Microsoft Corporation and others. Licensed under the MIT license.
 // SPDX-License-Identifier: MIT
 
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { getCocoaPodsSearch } from '../api/clearlyDefined'
+import { getCrateSearch } from '../../../api/clearlyDefined'
 import { AsyncTypeahead } from 'react-bootstrap-typeahead'
-import searchSvg from '../images/icons/searchSvg.svg'
-import 'react-bootstrap-typeahead/css/Typeahead.css'
+import searchSvg from '../../../images/icons/searchSvg.svg'
 
-export default class CocoaPodsSelector extends Component {
+export default class CrateSelector extends Component {
   static propTypes = {
     onChange: PropTypes.func
   }
 
   constructor(props) {
     super(props)
-    this.state = { isLoading: false, options: [], focus: true }
+    this.state = { isLoading: false, options: [], focus: false }
     this.getOptions = this.getOptions.bind(this)
     this.onChange = this.onChange.bind(this)
   }
@@ -23,13 +22,13 @@ export default class CocoaPodsSelector extends Component {
   onChange(values) {
     const { onChange } = this.props
     const value = values.length === 0 ? null : values[0]
-    value && onChange && onChange({ type: 'pod', provider: 'cocoapods', name: value.id }, 'package')
+    value && onChange && onChange({ type: 'crate', provider: 'cratesio', name: value.id }, 'package')
   }
 
   async getOptions(value) {
     try {
       this.setState({ ...this.state, isLoading: true })
-      const options = await getCocoaPodsSearch(this.props.token, value)
+      const options = await getCrateSearch(this.props.token, value)
       this.setState({ ...this.state, options, isLoading: false })
     } catch (error) {
       this.setState({ ...this.state, options: [], isLoading: false })
@@ -44,11 +43,11 @@ export default class CocoaPodsSelector extends Component {
           <img src={searchSvg} alt="search" />
         </div>
         <AsyncTypeahead
-          id="pod-selector"
+          id="crate-selector"
           className="harvest-search"
           useCache={false}
           options={options}
-          placeholder={'Pick a CocoaPod to harvest'}
+          placeholder={'Pick a Crate to harvest'}
           onChange={this.onChange}
           labelKey="id"
           onFocus={() => this.setState({ ...this.state, focus: true })}
