+++ src/components/Providers/Selectors/NpmSelector.js
@@ -3,11 +3,12 @@
 
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { getNugetSearch } from '../api/clearlyDefined'
+import { getNpmSearch } from '../../../api/clearlyDefined'
 import { AsyncTypeahead } from 'react-bootstrap-typeahead'
-import searchSvg from '../images/icons/searchSvg.svg'
+import searchSvg from '../../../images/icons/searchSvg.svg'
+import 'react-bootstrap-typeahead/css/Typeahead.css'
 
-export default class NuGetSelector extends Component {
+export default class NpmSelector extends Component {
   static propTypes = {
     onChange: PropTypes.func
   }
@@ -22,13 +23,13 @@ export default class NuGetSelector extends Component {
   onChange(values) {
     const { onChange } = this.props
     const value = values.length === 0 ? null : values[0]
-    value && onChange && onChange({ type: 'nuget', provider: 'nuget', name: value.id }, 'package')
+    value && onChange && onChange({ type: 'npm', provider: 'npmjs', name: value.id }, 'package')
   }
 
   async getOptions(value) {
     try {
       this.setState({ ...this.state, isLoading: true })
-      const options = await getNugetSearch(this.props.token, value)
+      const options = await getNpmSearch(this.props.token, value)
       this.setState({ ...this.state, options, isLoading: false })
     } catch (error) {
       this.setState({ ...this.state, options: [], isLoading: false })
@@ -43,16 +44,17 @@ export default class NuGetSelector extends Component {
           <img src={searchSvg} alt="search" />
         </div>
         <AsyncTypeahead
-          id="nuget-selector"
+          id="npm-selector"
           className="harvest-search"
+          inputProps={{ dataTestId: 'npm-selector' }}
           useCache={false}
           options={options}
-          placeholder={'Pick a Nuget to harvest'}
+          placeholder={'Select an NPM component'}
           onChange={this.onChange}
-          labelKey="id"
-          clearButton
           onFocus={() => this.setState({ ...this.state, focus: true })}
           onBlur={() => this.setState({ ...this.state, focus: false })}
+          labelKey="id"
+          clearButton
           highlightOnlyResult
           emptyLabel=""
           selectHintOnEnter
