+++ src/components/Providers/Selectors/PyPiSelector.js
@@ -3,12 +3,12 @@
 
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { getNpmSearch } from '../api/clearlyDefined'
+import { getPyPiSearch } from '../../../api/clearlyDefined'
 import { AsyncTypeahead } from 'react-bootstrap-typeahead'
-import searchSvg from '../images/icons/searchSvg.svg'
+import searchSvg from '../../../images/icons/searchSvg.svg'
 import 'react-bootstrap-typeahead/css/Typeahead.css'
 
-export default class NpmSelector extends Component {
+export default class PyPiSelector extends Component {
   static propTypes = {
     onChange: PropTypes.func
   }
@@ -23,13 +23,13 @@ export default class NpmSelector extends Component {
   onChange(values) {
     const { onChange } = this.props
     const value = values.length === 0 ? null : values[0]
-    value && onChange && onChange({ type: 'npm', provider: 'npmjs', name: value.id }, 'package')
+    value && onChange && onChange({ type: 'pypi', provider: 'pypi', name: value.id }, 'package')
   }
 
   async getOptions(value) {
     try {
       this.setState({ ...this.state, isLoading: true })
-      const options = await getNpmSearch(this.props.token, value)
+      const options = await getPyPiSearch(this.props.token, value)
       this.setState({ ...this.state, options, isLoading: false })
     } catch (error) {
       this.setState({ ...this.state, options: [], isLoading: false })
@@ -44,16 +44,15 @@ export default class NpmSelector extends Component {
           <img src={searchSvg} alt="search" />
         </div>
         <AsyncTypeahead
-          id="npm-selector"
+          id="pypi-selector"
           className="harvest-search"
-          inputProps={{ dataTestId: 'npm-selector' }}
           useCache={false}
           options={options}
-          placeholder={'Select an NPM component'}
+          placeholder={'Pick a PyPi to harvest'}
           onChange={this.onChange}
+          labelKey="id"
           onFocus={() => this.setState({ ...this.state, focus: true })}
           onBlur={() => this.setState({ ...this.state, focus: false })}
-          labelKey="id"
           clearButton
           highlightOnlyResult
           emptyLabel=""
