+++ src/components/Providers/Selectors/RubyGemsSelector.js
@@ -3,18 +3,19 @@
 
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { getCrateSearch } from '../api/clearlyDefined'
+import { getRubyGemsSearch } from '../../../api/clearlyDefined'
 import { AsyncTypeahead } from 'react-bootstrap-typeahead'
-import searchSvg from '../images/icons/searchSvg.svg'
+import searchSvg from '../../../images/icons/searchSvg.svg'
+import 'react-bootstrap-typeahead/css/Typeahead.css'
 
-export default class CrateSelector extends Component {
+export default class RubyGemsSelector extends Component {
   static propTypes = {
     onChange: PropTypes.func
   }
 
   constructor(props) {
     super(props)
-    this.state = { isLoading: false, options: [], focus: false }
+    this.state = { isLoading: false, options: [], focus: true }
     this.getOptions = this.getOptions.bind(this)
     this.onChange = this.onChange.bind(this)
   }
@@ -22,13 +23,13 @@ export default class CrateSelector extends Component {
   onChange(values) {
     const { onChange } = this.props
     const value = values.length === 0 ? null : values[0]
-    value && onChange && onChange({ type: 'crate', provider: 'cratesio', name: value.id }, 'package')
+    value && onChange && onChange({ type: 'gem', provider: 'rubygems', name: value.id }, 'package')
   }
 
   async getOptions(value) {
     try {
       this.setState({ ...this.state, isLoading: true })
-      const options = await getCrateSearch(this.props.token, value)
+      const options = await getRubyGemsSearch(this.props.token, value)
       this.setState({ ...this.state, options, isLoading: false })
     } catch (error) {
       this.setState({ ...this.state, options: [], isLoading: false })
@@ -43,11 +44,11 @@ export default class CrateSelector extends Component {
           <img src={searchSvg} alt="search" />
         </div>
         <AsyncTypeahead
-          id="crate-selector"
+          id="ruby-selector"
           className="harvest-search"
           useCache={false}
           options={options}
-          placeholder={'Pick a Crate to harvest'}
+          placeholder={'Pick a RubyGem to harvest'}
           onChange={this.onChange}
           labelKey="id"
           onFocus={() => this.setState({ ...this.state, focus: true })}
