+++ src/components/Providers/VersionPickers/RubyGemsVersionPicker.js
@@ -3,11 +3,11 @@
 
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { getCrateRevisions } from '../api/clearlyDefined'
-import Autocomplete from './Navigation/Ui/Autocomplete'
-import searchSvg from '../images/icons/searchSvg.svg'
+import { getRubyGemsRevisions } from '../../../api/clearlyDefined'
+import Autocomplete from '../../Navigation/Ui/Autocomplete'
+import searchSvg from '../../../images/icons/searchSvg.svg'
 
-export default class CrateVersionPicker extends Component {
+export default class RubyGemsVersionPicker extends Component {
   static propTypes = {
     onChange: PropTypes.func,
     request: PropTypes.object.isRequired,
@@ -16,7 +16,12 @@ export default class CrateVersionPicker extends Component {
 
   constructor(props) {
     super(props)
-    this.state = { customValues: [], options: [], focus: false }
+    this.state = {
+      customValues: [],
+      options: [],
+      focus: false,
+      selected: props.request.revision ? [props.request.revision] : []
+    }
     this.onChange = this.onChange.bind(this)
     this.filter = this.filter.bind(this)
   }
@@ -25,10 +30,14 @@ export default class CrateVersionPicker extends Component {
     this.getOptions('')
   }
 
+  componentWillReceiveProps(nextProps) {
+    this.setState({ ...this.state, selected: nextProps.request.revision ? [nextProps.request.revision] : [] })
+  }
+
   async getOptions(value) {
     try {
       const { name } = this.props.request
-      const options = await getCrateRevisions(this.props.token, name)
+      const options = await getRubyGemsRevisions(this.props.token, name)
       this.setState({ ...this.state, options })
     } catch (error) {
       this.setState({ ...this.state, options: [] })
@@ -54,7 +63,7 @@ export default class CrateVersionPicker extends Component {
 
   render() {
     const { defaultInputValue } = this.props
-    const { customValues, options, focus } = this.state
+    const { customValues, options, selected, focus } = this.state
     const list = customValues.concat(options)
     return (
       <div className={`harvest-searchbar ${focus ? 'active' : ''}`}>
@@ -62,19 +71,22 @@ export default class CrateVersionPicker extends Component {
           <img src={searchSvg} alt="search" />
         </div>{' '}
         <Autocomplete
-          id="crate-version-picker"
+          id="rubygems-version-picker"
+          selected={selected}
           options={list}
           defaultInputValue={defaultInputValue}
-          placeholder={options.length === 0 ? 'Could not fetch versions, type a Crate version' : 'Pick a Crate version'}
+          placeholder={
+            options.length === 0 ? 'Could not fetch versions, type a RubyGem version' : 'Pick a RubyGem version'
+          }
           onChange={this.onChange}
-          positionFixed
-          clearButton
           onFocus={() => this.setState({ ...this.state, focus: true })}
           onBlur={() => this.setState({ ...this.state, focus: false })}
+          positionFixed
+          clearButton
           allowNew
           newSelectionPrefix="Version:"
           emptyLabel=""
-          // filterBy={this.filter}
+          filterBy={this.filter}
           selectHintOnEnter
         />
       </div>
